Task 2 commit

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -7,9 +7,9 @@
  * Return: the length of the string
  */
 
-int _strlen(char *s)
+size_t _strlen(char *s)
 {
-	int i = 0;
+	size_t i = 0;
 
 	while (s[i] != '\0')
 		i++;
